+++ src-tauri/src/compress.rs
@@ -47,10 +47,11 @@ pub enum FileEntryStatus {
 }
 
 #[allow(clippy::upper_case_acronyms)]
-#[derive(Debug, Eq, PartialEq, Clone, serde::Serialize, serde::Deserialize, Type)]
+#[derive(Debug, Eq, PartialEq, Clone, serde::Serialize, serde::Deserialize, Type, Default)]
 pub enum ImageType {
     JPEG,
     PNG,
+    #[default]
     WEBP,
     GIF,
     TIFF,